+++ src/main/java/ru/skypro/homework/service/impl/CommentServiceImpl.java
@@ -3,6 +3,8 @@ package ru.skypro.homework.service.impl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.skypro.homework.dto.CommentDTO;
+import ru.skypro.homework.dto.Role;
+import ru.skypro.homework.exception.UserNotRegisteredException;
 import ru.skypro.homework.model.Ads;
 import ru.skypro.homework.model.Comment;
 import ru.skypro.homework.model.UserInfo;
@@ -13,6 +15,7 @@ import ru.skypro.homework.service.mapper.CommentMapper;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 @Service
@@ -27,42 +30,52 @@ public class CommentServiceImpl implements CommentService {
     public List<CommentDTO> getCommentsByAdsId(long id) {
         Ads ads = adsRepository.findById(id).orElse(null);
         List<Comment> commentList = ads.getComments();
-        List<CommentDTO> commentDTOList = new ArrayList<>();
-        for (Comment comment:commentList){
-            commentDTOList.add(commentMapper.commentToCommentDto(comment));
+        List<CommentDTO> commentDTOList = Collections.singletonList
+                (commentMapper.commentToCommentDto((Comment) commentList));
+        for (int i = 0; i < commentList.size(); i++) {
+            if (commentList.get(i).getAuthor().getImageModel() != null) {
+                commentDTOList.get(i).setAuthorImage
+                        ("/users/image/" + commentList.get(i).getAuthor().getImageModel().getId());
+            }
         }
+
         return commentDTOList;
     }
 
     @Override
     public CommentDTO addComment(long adsId, String text) {
         UserInfo user = authService.getCurrentUser();
-        Ads ads = adsRepository.findById(adsId).orElse(null);
-        Comment comment = new Comment();
-        comment.setText(text);
-        comment.setAds(ads);
-        comment.setAuthor(user);
-        comment.setCreatedAt(String.valueOf(LocalDateTime.now()));;
-        commentRepository.save(comment);
-
-        return commentMapper.commentToCommentDto(comment);
+        if (user == null) {
+            throw new UserNotRegisteredException("The user is not registered");
+        } else {
+            Ads ads = adsRepository.findById(adsId).orElse(null);
+            Comment comment = new Comment();
+            comment.setText(text);
+            comment.setAds(ads);
+            comment.setAuthor(user);
+            commentRepository.save(comment);
+            return commentMapper.commentToCommentDto(comment);
+        }
     }
 
     @Override
     public void deleteComment(long adsId, long commentId) {
-
-        commentRepository.deleteById(commentId);
-
+        UserInfo user = authService.getCurrentUser();
+        Ads ads = adsRepository.findById(adsId).orElse(null);
+        if (user.getId() == ads.getAuthor().getId() || user.getRole().equals(Role.ADMIN)) {
+            commentRepository.deleteById(commentId);
+        }
     }
 
     @Override
     public CommentDTO updateComment(long adsId, long commentId, String text) {
-
+        UserInfo user = authService.getCurrentUser();
+        Ads ads = adsRepository.findById(adsId).orElse(null);
         Comment comment = commentRepository.findById(commentId).orElse(null);
-        comment.setText(text);
-        comment.setCreatedAt(String.valueOf(LocalDateTime.now()));;
-        commentRepository.save(comment);
-
+        if (user.getId() == ads.getAuthor().getId() || user.getRole().equals(Role.ADMIN)) {
+            comment.setText(text);
+            commentRepository.save(comment);
+        }
         return commentMapper.commentToCommentDto(comment);
     }
 }
